requirements.txt
google-cloud-aiplatform 
langchain-google-vertexai 
langgraph 
vertexai
pydantic
PyPDF2
python-docx
pydantic_settings
